Removed test_klp_util.py
